OPENNLP-1351: Moving the tokenize function.
